backend/requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.7
django-extensions==2.1.0
djangorestframework==3.8.2
pkg-resources==0.0.0
python-decouple==3.1
pytz==2018.5
six==1.11.0
static3==0.7.0
